% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorCov.R
\name{CorCov}
\alias{CorCov}
\title{Convert correlation to covariance}
\usage{
CorCov(sd1, sd2, rho)
}
\arguments{
\item{sd1}{Standard deviation of the first variable (numeric).}

\item{sd2}{Standard deviation of the second variable (numeric).}

\item{rho}{Correlation between the two variables (numeric).}
}
\value{
Covariance (numeric).
}
\description{
Computes the covariance between two variables given their standard deviations and correlation.
}
\examples{
CorCov(2, 3, 0.5)  # returns 3
}
