% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimpmulti_methods.R
\name{dataimp-multi-methods}
\alias{dataimp-multi-methods}
\alias{print.dataimp_multi}
\alias{summary.dataimp_multi}
\alias{print.summary.dataimp_multi}
\title{Methods for dataimp_multi objects}
\usage{
\method{print}{dataimp_multi}(x, ...)

\method{summary}{dataimp_multi}(object, ...)

\method{print}{summary.dataimp_multi}(x, ...)
}
\arguments{
\item{x, object}{An object of class \code{dataimp_multi}.}

\item{...}{Further arguments passed to or from other methods (currently ignored).}
}
\value{
- \code{print.dataimp_multi}: prints a short description of the object.
- \code{summary.dataimp_multi}: returns an object of class \code{summary.dataimp_multi}.
- \code{print.summary.dataimp_multi}: prints the summary.
}
\description{
Methods for \code{dataimp_multi} objects
}
\details{
Printing and summarizing methods for objects returned by \code{\link{genimp_multi}}.
}
\examples{
imps <- list(
  function(n) rnorm(n, 0, 1),
  function(n) rnorm(n, 1, 1),
  function(n) rnorm(n, 2, 1)
)
df <- data.frame(
  eff1 = c(0.5, NA, 0.3),
  eff2 = c(NA, 0.2, 0.3),
  eff3 = c(0.1, 0.2, NA),
  se1  = c(0.1, NA, 0.2),
  se2  = c(NA, 0.25, 0.2),
  se3  = c(0.05, 0.1, NA),
  cor12 = c(NA, 0.4, 0.5),
  cor13 = c(0.3, NA, 0.6),
  cor23 = c(NA, 0.3, 0.7),
  N = c(50, 60, 55)
)

out <- genimp_multi(df, iter = 5, imps = imps,
                    effs = c("eff1","eff2","eff3"),
                    ses = c("se1","se2","se3"),
                    cors = c("cor12","cor13","cor23"),
                    Ns = "N", imprho = c(0.7,0.5,0.6))
summary(out)

}
