% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genimp_methods.R
\name{genimp-methods}
\alias{genimp-methods}
\alias{print.genimp}
\alias{summary.genimp}
\alias{print.summary.genimp}
\alias{plot.genimp}
\title{Methods for \code{genimp} objects}
\usage{
\method{print}{genimp}(x, ...)

\method{summary}{genimp}(object, ...)

\method{print}{summary.genimp}(x, ...)

\method{plot}{genimp}(x, ...)
}
\arguments{
\item{x, object}{An object of class \code{genimp}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
- \code{print.genimp}: prints a short description of the imputation object.
- \code{summary.genimp}: returns an object of class \code{summary.dataimp}.
- \code{print.summary.genimp}: prints the summary.
- \code{plot.genimp}: creates a basic plot (e.g., distribution of imputed effects).

#' @details
When \code{genimp()} is called repeatedly (e.g. inside \code{lapply} or \code{mapply}),
the result will be a list of \code{genimp} objects. In that case, use
\code{lapply(out, summary)} or \code{purrr::map(out, summary)} to obtain summaries
for each element.
}
\description{
These methods provide printing, summarizing, and plotting facilities for
objects of class \code{genimp}, created by \code{\link{genimp}}.
}
\examples{
dmnar <- data.frame(
  EstCR = c(3.7, 3.5, 4.2, NA),
  EstSR = c(3.2, 3.2, NA, 5.6),
  SECR = c(1.5, 1.3, 2.1, NA),
  SESR = c(0.7, 1.8, NA, 2.2),
  Cor.ws = c(0.75, 0.5, NA, NA),
  N = c(100, 150, 120, 95)
)
imp1 <- function(n) rnorm(n, mean = 3, sd = 2)
imp2 <- function(n) runif(n, min = 4, max = 7)
out <- genimp(df = dmnar, iter = 2,
              imp1 = imp1, imp2 = imp2,
              eff1 = "EstCR", eff2 = "EstSR",
              se1 = "SECR", se2 = "SESR",
              cor = "Cor.ws", N = "N")

out              # calls print.genimp
summary(out)     # calls summary.genimp
plot(out)        # calls plot.genimp

}
