% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumeth_methods.R
\name{sumeth-methods}
\alias{sumeth-methods}
\alias{print.sumeth}
\alias{summary.sumeth}
\alias{print.summary.sumeth}
\alias{plot.sumeth}
\title{Methods for \code{sumeth} objects}
\usage{
\method{print}{sumeth}(x, ...)

\method{summary}{sumeth}(object, ...)

\method{print}{summary.sumeth}(x, ...)

\method{plot}{sumeth}(x, ...)
}
\arguments{
\item{x, object}{An object of class \code{sumeth}.}

\item{...}{Further arguments passed to or from other methods (currently ignored).}
}
\value{
- \code{print.sumeth}: prints a compact summary table of the pooled results.
- \code{summary.sumeth}: returns an object of class \code{summary.sumeth}.
- \code{print.summary.sumeth}: prints the detailed summary including degrees of freedom.
}
\description{
Provides printing and summarizing for objects returned by \code{\link{sumeth}}.
}
\examples{
res <- data.frame(
  eff1 = c(3.7, 3.5),
  eff2 = c(3.2, 3.2),
  se1 = c(1.5, 1.3),
  se2 = c(2.7, 1.8),
  cov12 = c(0.02, 0.5)
)
out <- sumeth(
  eff1 = res$eff1, eff2 = res$eff2,
  se1 = res$se1, se2 = res$se2,
  cov12 = res$cov12, method = "Normal (1, 6)"
)
out
summary(out)

}
