% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepooldata.R
\name{makepooldata}
\alias{makepooldata}
\title{Helper to create Q_mat and U_list for poolsums}
\usage{
makepooldata(data, effs = "eff", ses = "se", covs = "cov")
}
\arguments{
\item{data}{Dataframe with effect sizes (named eff1, eff2, ...), standard errors (named se1, se2, ...),
covariances (named cov12, cov13,...).}

\item{effs}{Prefix for effect size columns (e.g., "eff")}

\item{ses}{Prefix for standard error columns (e.g., "se")}

\item{covs}{Prefix for covariance columns (e.g., "cov")}
}
\value{
A list with Q_mat and U_list
}
\description{
Helper to create Q_mat and U_list for poolsums
}
